py: Further refactoring of docker client
